Move login/logout buttons to header component.
